Move ActorOriginValidator from membership to common
