+++ runtime-modules/common/src/lib.rs
@@ -2,3 +2,4 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 pub mod currency;
+pub mod origin_validator;
